test: restore Groovy spec helpers after rebase
